# src/routes/inventoryRoutes.ts
/**
 * 庫存管理 API 路由
 * 路徑: /api/v1/refrigerators/:refrigeratorId/inventory
 */

import { Router, Request, Response } from 'express';
import type { 
  CreateInventoryInput, 
  UpdateInventoryInput, 
  ConsumeInventoryInput,
  InventoryPaginationParams,
  UpdateInventorySettingsInput,
} from '../types/inventory.js';
import * as inventoryService from '../services/inventoryService.js';

const router = Router({ mergeParams: true });

// ===== Helper: 取得 userId 和 refrigeratorId =====
const getUserId = (req: Request): string | null => {
  return (req.headers['x-user-id'] as string) ?? null;
};

const getRefrigeratorId = (req: Request): string => {
  return req.params.refrigeratorId;
};

// ===== 3.1 取得庫存列表 =====
router.get('/', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const params: InventoryPaginationParams = {
      page: req.query.page ? parseInt(req.query.page as string, 10) : 1,
      limit: req.query.limit ? parseInt(req.query.limit as string, 10) : 50,
      category: req.query.category as string | undefined,
      status: req.query.status as InventoryPaginationParams['status'],
      include: req.query.include as string | undefined,
    };

    const result = await inventoryService.getInventoryItems(userId, refrigeratorId, params);
    
    res.json({
      status: true,
      data: {
        items: result.items,
        total: result.total,
        ...(result.stats && { stats: result.stats }),
        ...(result.summary && { summary: result.summary }),
      },
    });
  } catch (error) {
    console.error('[Inventory] 取得列表失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.8 取得分類列表 =====
router.get('/categories', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const categories = await inventoryService.getInventoryCategories(userId, refrigeratorId);
    res.json({ status: true, data: { categories } });
  } catch (error) {
    console.error('[Inventory] 取得分類失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.9 取得庫存摘要 =====
router.get('/summary', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const summary = await inventoryService.getInventorySummary(userId, refrigeratorId);
    res.json({ status: true, data: { summary } });
  } catch (error) {
    console.error('[Inventory] 取得摘要失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.11 取得庫存設定 =====
router.get('/settings', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const settings = await inventoryService.getInventorySettings(userId, refrigeratorId);
    res.json({ status: true, data: { settings } });
  } catch (error) {
    console.error('[Inventory] 取得設定失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.11 更新庫存設定 (PUT) =====
router.put('/settings', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const input = req.body as UpdateInventorySettingsInput;
    const settings = await inventoryService.updateInventorySettings(userId, refrigeratorId, input);
    
    res.json({
      status: true,
      message: '設定已更新',
      data: { settings },
    });
  } catch (error) {
    console.error('[Inventory] 更新設定失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.11 部分更新庫存設定 (PATCH) =====
router.patch('/settings', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const input = req.body as UpdateInventorySettingsInput;
    const settings = await inventoryService.updateInventorySettings(userId, refrigeratorId, input);
    
    res.json({
      status: true,
      message: '設定已更新',
      data: { settings },
    });
  } catch (error) {
    console.error('[Inventory] 更新設定失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.2 取得單筆食材 =====
router.get('/:id', async (req: Request, res: Response) => {
  try {
    const refrigeratorId = getRefrigeratorId(req);
    const { id } = req.params;

    const item = await inventoryService.getInventoryById(id, refrigeratorId);
    
    if (!item) {
      res.status(404).json({ status: false, error: '找不到該食材' });
      return;
    }

    res.json({ status: true, data: { item } });
  } catch (error) {
    console.error('[Inventory] 取得單筆失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.3 新增食材 =====
router.post('/', async (req: Request, res: Response) => {
  try {
    const userId = getUserId(req);
    const refrigeratorId = getRefrigeratorId(req);

    if (!userId) {
      res.status(400).json({ status: false, error: '缺少 X-User-Id' });
      return;
    }

    const input = req.body as CreateInventoryInput;
    
    if (!input.name) {
      res.status(400).json({ status: false, error: '缺少食材名稱' });
      return;
    }

    const item = await inventoryService.createInventoryItem(userId, refrigeratorId, input);
    
    res.status(201).json({
      status: true,
      message: 'Created successfully',
      data: { id: item.id },
    });
  } catch (error) {
    console.error('[Inventory] 新增失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.4 更新食材 =====
router.put('/:id', async (req: Request, res: Response) => {
  try {
    const refrigeratorId = getRefrigeratorId(req);
    const { id } = req.params;
    const input = req.body as UpdateInventoryInput;

    const item = await inventoryService.updateInventoryItem(id, refrigeratorId, input);
    
    if (!item) {
      res.status(404).json({ status: false, error: '找不到該食材' });
      return;
    }

    res.json({
      status: true,
      message: 'Updated successfully',
      data: { id: item.id },
    });
  } catch (error) {
    console.error('[Inventory] 更新失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.5 刪除食材 =====
router.delete('/:id', async (req: Request, res: Response) => {
  try {
    const refrigeratorId = getRefrigeratorId(req);
    const { id } = req.params;

    const deleted = await inventoryService.deleteInventoryItem(id, refrigeratorId);
    
    if (!deleted) {
      res.status(404).json({ status: false, error: '找不到該食材' });
      return;
    }

    res.json({ status: true, message: 'Deleted successfully' });
  } catch (error) {
    console.error('[Inventory] 刪除失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

// ===== 3.12 消耗食材 =====
router.post('/:id/consume', async (req: Request, res: Response) => {
  try {
    const refrigeratorId = getRefrigeratorId(req);
    const { id } = req.params;
    const input = req.body as ConsumeInventoryInput;

    if (!input.quantity || input.quantity <= 0) {
      res.status(400).json({ 
        status: false, 
        code: 'INV_001',
        error: '缺少或無效的 quantity 欄位',
      });
      return;
    }

    if (!input.reasons || input.reasons.length === 0) {
      res.status(400).json({ 
        status: false, 
        error: '缺少 reasons 欄位',
      });
      return;
    }

    const result = await inventoryService.consumeInventoryItem(id, refrigeratorId, input);
    
    if (!result) {
      res.status(404).json({ 
        status: false, 
        code: 'INV_004',
        error: '找不到該食材',
      });
      return;
    }

    res.json({
      status: true,
      message: 'Consumed successfully',
      data: result,
    });
  } catch (error) {
    if (error instanceof Error && error.message === '消耗數量超過庫存') {
      res.status(400).json({ 
        status: false, 
        code: 'INV_002',
        error: error.message,
      });
      return;
    }
    console.error('[Inventory] 消耗失敗:', error);
    res.status(500).json({ status: false, error: '伺服器錯誤' });
  }
});

export default router;